Creating a class to accomodate multiple attachments in MIME standard for SES emails
